+++ src/components/Header.css
@@ -9,7 +9,6 @@ header {
   background-color: #0018a4;
   font-size: 16px;
   width: 100%;
-
 }
 
 h1 {
