+++ src/components/Header.jsx
@@ -1,9 +1,9 @@
-import "./Header.css"
+import "./Header.css";
 
 export const Header = () => {
   return (
     <header>
       <h1>Hogwarts Survey</h1>
     </header>
-  )
-}
+  );
+};
