+++ src/components/Submit.jsx
@@ -1,0 +1,7 @@
+export const Submit = () => {
+  return (
+    <button className="submit-button" type="submit">
+      Submit
+    </button>
+  );
+};
