+++ src/components/Summary.jsx
@@ -1,13 +1,13 @@
-// import { useState } from "react";
-
-// export const Summary = () => {
-//   const [summary, setSummary] = useState()
-
-//   const handleSubmit = () => {
-
-
-//   }
-//   return (
-//     <button>Submit</button>
-//   )
-// }
+export const Summary = ({ houseName, characterName, spellName, clearForm }) => {
+  return (
+    <div className="results">
+      <h3>Here are your results!</h3>
+      <p>Your favorite character is {characterName}.</p>
+      <p>Your favorite spell is {spellName}.</p>
+      <p>You would like to be in house {houseName}!</p>
+      <button className="clear-button" onClick={clearForm}>
+        Again!
+      </button>
+    </div>
+  );
+};
